Initial hack at code gen with Roslyn
Add the .NET Compiler Platform (Roslyn) NuGet package, and experiment a bit with using it to generate code from scratch, which we'll have to do.

This uncovered a problem with the Roslyn NuGet packages, explained in this forum post:

https://social.msdn.microsoft.com/Forums/vstudio/en-US/c03febde-cf6d-478e-ac01-99102eb07f1b/the-language-c-is-not-supported-bug-on-newcompilationunitformat?forum=roslyn

As a workaround, we introduce the targets file CopyRoslynImplementationAssemblies.targets, and import it into the project files for DataModelGenerator and JSchema.Tests.