Fix #20: Json.NET can't deserialize a property named "$ref"
To work around this problem, we set the JsonProperty attribute of the
JsonSchema.Reference property to "$$ref". (We did this in a previous
commit.) Then we use a regex to change "$ref" to "$$ref" after we read
the schema file and before we ask Json.NET to deserialize it into a
JsonSchema object. The regex accounts for the quotes surrounding the
property name and the colon that follows it.