Partial #17: Give object-valued properties the correct type from the dictionary
Unit tests pass, but:
The program fails with null-ref because we don't keep track of the
outermost schema, where the definitions are (and for that matter we
don't really support definitions in inner schemas).
We need unit tests for the InferType logic.
We don't support arrays of arrays -- InferredType should encapsulate that.