Fix #17: get nested definitions to work
This completes the work on #17, "Give object-valued properties the correct
type from the dictionary".

There was a NullReferenceException when a type defined in one definition
referred to a type defined in another definition. The fix is to carry
around the root schema (where the definitions are), because at the moment
we don't support definitions in subschemas (we assume they are all
located in the root schema), because at the moment we don't support
JPointer: we just support $ref values of the form "#/definitions/typeName".