Fix #52: Avoid arbitrary code execution when deserializing HintDictionary
As originally designed, the code hint mechanism allowed the user to
specify an arbitrary CodeGenHint-derived type in the hints file. This was
convenient because it allowed JSchema to use Json.NET to directly
deserialize the hints, but it potentially allowed for arbitrary code
execution as the type was being deserialized.

So JSchema now uses a mechanism where there is a fixed set of hint types.
The hints file contains an enumeration value to specify the hint type, and
a property bag used to initialize the hint object. JSchema has explicit
code to initialize each type of hint from the property bag.