Sarif: Fix NuGet package build break
The Json.Schema NuGet file is built from the JsonSchemaValidator project because, in addition to containing Json.Schema.dll, it also contains JsonSchemaValidator.exe, so we can't build the package until both the .dll and the .exe are built.

The package creation started failing, complaining that it can't find System.Collections.Immutable.dll.  The .nuspec package does indeed ask to place Immutable.dll in the tools directory. And the tool does not depend on Immutable.dll, so it is indeed not in the output directory. So I understand why it fails -- but not why it just started failing.

Removed file reference to Immutable.dll from the .nuspec.