Build Json.Pointer for netstandard2.0
This is the first in a series of commits to change all of
JSchema to .netstandard and dotnet core.

We start with Json.Pointer. We create a new versions of
Everything.sln that contains just those projects we've
converted to netstandard. We create a new build.props
that takes full advantage of the properties defined in the
SDK target files. We don't need AssemblyInfo.cs any more.