Generate System.Uri-valued property for uri-reference format
When the code generator generates a .NET property from a schema
property whose type is `string`, the .NET property also has type
`string` -- _unless_ the schema property has format `uri` (in
which case the .NET property has type `System.Uri`) or `date-time`
(in which case the .NET property has type `System.DateTime`).

But the code generator didn't take account of the `uri-reference`
format, and it generated a `string` property when it should have
generated a `System.Uri`. This change fixes that problem.