Fix #17: Code generation fails for an array of uri-formatted strings (#18)
* Fix #17: Code generation fails for an array of uri-formatted strings

The cloning code that copies one collection to another is generated differently depending on the collection's element type. We did not cover the case where the element type was a URI, so it defaulted to treating the element type as a collection type, which led to a crash.

It's a one-line fix with an additional unit test.