Remove `ExtensionCapturingTraceWriter` for forward compatibility with recent Newtonsoft.Json versions (#123)
For the sake of compatibility with consumers who both (1) use JSchema (which depends on Newtonsoft.Json 9.0.1) and (2) use other packages which depend on Newtonsoft.Json 11 or later, we remove the internal `ExtensionCapturingTraceWriter` class. There is no effect on observed behavior. We replace the functionality of `ExtensionCapturingTraceWriter` with a new class `SchemaValidationErrorAccumulator`.

This change is related to a long-standing problem in Newtonsoft.Json (see https://github.com/JamesNK/Newtonsoft.Json/issues/1616). Version 11 fixed that problem by removing the type `Newtonsoft.Json.TraceLevel` entirely (that is, now, both the net461 and netstandard2.0 libraries use `System.Diagnostics.TraceLevel`). But JSchema, relying as it does on 9.0.1, _requires_ `Netwonsoft.Json.TraceLevel` to implement an `ITraceWriter`.

In an application with the dependencies described above, the higher version of Newtonsoft.Json will be loaded, the type `Newtonsoft.Json.TraceLevel` will not be available, and so the type initializer for `ExtensionCapturingTraceWriter` will fail.